where applicable read details from ${projectDir}/build/LCP.json otherwise use ${projectDir}/LCP.json

+++ scripts/k8s/delete.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
-import ytt
 import os
-import uuid
 
-tmpfile = "/tmp/delete-%s.yaml" % uuid.uuid1()
+name = os.environ.get("NAME")
 
-with open(tmpfile, "w") as f:
-    f.write(ytt.generate_workload_yaml())
-os.popen("kubectl delete --ignore-not-found --wait -f %s" % tmpfile).read()
-os.remove(tmpfile)
+tmpfile = "/tmp/.%s.yaml" % name
+
+if os.path.exists(tmpfile):
+    os.popen("kubectl delete --ignore-not-found --wait -f %s" % tmpfile).read()
+    os.remove(tmpfile)
+    os.remove("/tmp/.%s.yttc" % name)
